+++ docs/changelog.md
@@ -3,6 +3,7 @@
 ## 0.9.14
 - Small tweak to the French (`"fr"`) localization.
 - Added Hebrew (`"he"`) localization (thank you @tinytim84!).
+- Added Thai (`"th"`) localization (thank you @samlaukinoon!).
 
 ## 0.9.13
 - Improvement to the French (`"fr"`) localization.